+++ src/interface/index.ts
@@ -1,3 +1,0 @@
-export * from './cli/create-todo/create-todo.interface';
-export * from './cli/cli.types';
-export * from './cli/create-todo/create-todo';