+++ src/main/app.ts
@@ -1,16 +1,20 @@
 import { Container } from './container';
-import { CliModule } from './modules';
+import { CliModule, Module, HttpModule } from './modules';
 
 export class App {
-  private cli_module: CliModule;
+  private cli_module: Module;
 
-  constructor({ cli = null }) {
+  private http_module: Module;
+
+  constructor({ cli = null, http = null }) {
     const container = new Container();
 
+    this.http_module = http || new HttpModule(container);
     this.cli_module = cli || new CliModule(container);
   }
 
   start(): void {
+    this.http_module.start();
     this.cli_module.start();
   }
 }