+++ src/main/modules/index.ts
@@ -1,2 +1,3 @@
 export * from './modules.interface';
 export * from './cli/cli';
+export * from './http/http';